Fixing imports after moving css files
